refactor(ic-certified-assets): flatten asset_certification/ to top-level modules
Replace the nested `asset_certification/{mod,tree,types/...}` hierarchy with
four top-level modules — `certification`, `http`, `nested_tree`, `rc_bytes` —
to remove an unnecessary level of nesting and put the certification primitives
in a single discoverable place.

The new `certification.rs` consolidates the path/hash types (AssetPath,
HashTreePath, NestedTreeKey, RequestHash, ResponseHash, CertificateExpression,
WitnessResult), the certified-response tree (CertifiedResponses), and the
helpers that produce response hashes and certificate expressions
(response_hash, build_ic_certificate_expression*). These were split between
`asset_certification/{mod.rs,types/certification.rs,types/http.rs}` for no
good reason — the helpers in http.rs were certification helpers parked next
to the HTTP types because both were in the same file.

`http.rs` now holds only HTTP-domain items: HttpRequest, HttpResponse and
its build_* helpers, StreamingCallback{Token,HttpResponse}, StreamingStrategy,
CallbackFunc, HeaderField, FALLBACK_FILE. The cycle through `state_machine`
remains for now — `HttpResponse::build_ok(&Asset, ...)` is addressed in the
next commit.

`rc_bytes.rs` and `nested_tree.rs` (renamed from `tree.rs`) are unchanged
content, promoted to the top level via `git mv`.
