--- ic-certified-assets/src/asset_certification/types/mod.rs
@@ -1,3 +1,0 @@
-pub mod certification;
-pub mod http;
-pub mod rc_bytes;
