+++ ic-certified-assets/src/nested_tree.rs
