+++ ic-certified-assets/src/rc_bytes.rs
